fix DataTypeVisitor to allows arrya of String
On service port, when a return value was somthing like String[], it was build a datatype as Array[Java.lang.string] due to

sb.append(name.substring(0, 1).toUpperCase)
sb.append(name.substring(1).toLowerCase)

on array datatype visitor.

replace this code by a visitor on the component datatype of the array.